backend/requirements.txt
annotated-types==0.7.0
fastapi==0.111.1
greenlet==3.2.4
loguru==0.7.3
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-sqlalchemy==0.46b0
opentelemetry-sdk==1.25.0
pydantic==2.12.3
pydantic_core==2.41.4
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.3
